Move license-check to DomComponent (#26087)
